projects/email_finder_kiro/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
boto3==1.35.0
nltk==3.9.1
hypothesis==6.115.0
pytest==8.3.0
pydantic==2.9.0
python-multipart==0.0.12
jinja2==3.1.4
